Create store and connect store to the app
