+++ src/Redux/configureStore.js
@@ -1,27 +1,14 @@
 /* eslint-disable linebreak-style */
 /* eslint-disable quotes */
-const BASE_URL = "https://api.coinstats.app/public/v1/coins";
-const FETCH_COINS = "FETCH_COINS";
-const initialState = [];
+import { applyMiddleware, compose, createStore } from "redux";
+import thunk from "redux-thunk";
+import logger from "redux-logger";
+import CryptoReducer from "./cryptoReducer";
 
-export const FetchData = (data) => ({
-  type: FETCH_COINS,
-  payload: data,
-});
+const composeEnhancers = window.__REDUX_DEVTOOLS_EXTENSION_COMPOSE__ || compose;
+const store = createStore(
+  CryptoReducer,
+  composeEnhancers(applyMiddleware(thunk, logger)),
+);
 
-export const LoadCrypto = () => async (dispatch) => {
-  const response = await fetch(BASE_URL);
-  const { data } = await response.json();
-  dispatch(FetchData(data));
-};
-
-const CryptoReducer = (state = initialState, action) => {
-  switch (action.type) {
-    case FETCH_COINS:
-      return action.payload;
-    default:
-      return state;
-  }
-};
-
-export default CryptoReducer;
+export default store;
